src/main/java/kitchenpos/application/dto/OrderTableCreateRequest.java
package kitchenpos.application.dto;

import javax.validation.constraints.Min;

import kitchenpos.domain.entity.OrderTable;

public class OrderTableCreateRequest {
    @Min(0)
    private int numberOfGuests;
    private boolean empty;

    private OrderTableCreateRequest() {
    }

    public OrderTableCreateRequest(int numberOfGuests, boolean empty) {
        this.numberOfGuests = numberOfGuests;
        this.empty = empty;
    }

    public OrderTable toEntity() {
        return new OrderTable(null, null, numberOfGuests, empty);
    }

    public int getNumberOfGuests() {
        return numberOfGuests;
    }

    public boolean isEmpty() {
        return empty;
    }
}
